FIX: Repair the five examples that could not run at all
None of these could be run by anyone, let alone rendered for the gallery:

- superquad_with_gui imported from enthought.traits.api and friends, the
  package names retired around 2010, so it has raised ImportError ever since
- delaunay_graph used np.int, removed in NumPy 1.24
- mlab_3D_to_2D pinned matplotlib to the wx backend, which now raises under a
  running Qt loop; the pin dates from a wx/Mayavi interaction and nothing in
  the example needs a particular backend
- normal_flipping_stl resolved its default data file relative to the working
  directory rather than to itself, so it only worked if run from its own folder
- data/cube.stl is not an STL: it is the Python snippet the cube was pasted
  from, wrapped in cube_stl = """...""", so vtkSTLReader read it as a binary
  file, failed, and handed back an empty polydata

+++ examples/mayavi/advanced_visualization/delaunay_graph.py
@@ -216,8 +216,8 @@ if __name__ == '__main__':
     # nx.mst returns an edge generator
     edges = nx.minimum_spanning_tree(g).edges(data=True)
     start_idx, end_idx, _ = np.array(list(edges)).T
-    start_idx = start_idx.astype(np.int)
-    end_idx   = end_idx.astype(np.int)
+    start_idx = start_idx.astype(int)
+    end_idx   = end_idx.astype(int)
 
     # Plot this with Mayavi
     graph_plot(x, y, z, start_idx, end_idx,
